refactor: simplify reset password flow by extracting form logic into a separate component
